tasks/web-burmalda/server/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["a <a@a.a>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pytonconnect = "^0.3.2"
aiohttp = "^3.11.18"
pydantic-settings = "^2.9.1"
tonsdk = "^1.0.15"
sqlalchemy = "^2.0.40"
jinja2 = "^3.1.6"
databases = "^0.9.0"
asyncpg = "^0.30.0"
pyjwt = "^2.8.0"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"
